+++ src/openhems/__init__.py
@@ -1,12 +1,13 @@
-""" A Home Energy Management System based on Home-Assistant """
 """
+A Home Energy Management System based on Home-Assistant
+
 Module to easily launch customized OpenHEMS Application.
 """
 
-__version__ = "0.1.12"
-__display_version__ = __version__  # used for command line version
-
+from .modules import *
 from .main import OpenHEMSApplication, main
 from .server import OpenHEMSServer
 from .modules.energy_strategy.energy_strategy import LOOP_DELAY_VIRTUAL
-from .modules import *
+
+# __version__ = "0.1.12"
+# __display_version__ = __version__  # used for command line version