+++ src/openhems/modules/energy_strategy/offpeak_strategy.py
@@ -81,13 +81,12 @@ class OffPeakStrategy(EnergyStrategy):
 		if marginPower<0:
 			self.logger.info("Can't switch on devices: not enough power margin : %s", marginPower)
 			return True
-		for elem in self.getNodes():
-			if not elem.isOn():
-				if self.switchOnSchedulable(elem, cycleDuration, True):
-					# Do just one at each loop to check Network constraint
-					return True
-				ok = False
-		return ok
+		for elem in self.getNodes(True):
+			switchOn = self.switchOnSchedulable(elem.node, cycleDuration, True)
+			if switchOn and elem.changed(switchOn):
+				self.logger.info("Switch on just one device at each loop to ensure Network constraint.")
+				return True
+		return False
 
 	def updateNetwork(self, cycleDuration:int, allowSleep:bool, now=None) -> int:
 		"""
@@ -157,7 +156,7 @@ class OffPeakStrategy(EnergyStrategy):
 			# TODO: set this margin in configuration
 			if missingTime>peakTime:
 				self.logger.warning("Missing %d minutes to respect timeout.",
-					                    missingTime-peakTime)
+					                    round((missingTime-peakTime).seconds/60))
 		self.logger.debug("OffpeakStrategy.getMissingTime(%s) = %s", schedule, missingTime)
 		return missingTime
 
@@ -183,12 +182,12 @@ class OffPeakStrategy(EnergyStrategy):
 				start, end = onPeriod
 				if now>start:
 					if end>now:
-						if not elem.isOn() and self.switchOnSchedulable(elem, cycleDuration, True):
+						if self.switchOnSchedulable(elem, cycleDuration, True):
 							self.logger.info(
 							    "Switch on '%s' due to missing time on offpeak periods to respect constraints.",
 							    elem.id)
 					else:
-						if not self.switchOnSchedulable(elem, cycleDuration, False):
+						if not elem.switchOn(False):
 							# TODO : remove past periods : useless anymore
 							schedule.setStrategyCache(self.strategyId, onPeriods)
 
@@ -217,7 +216,8 @@ class OffPeakStrategy(EnergyStrategy):
 				peakPeriods.append([previousRangeEnd, availableTime, rangeEnd, cost])
 			previousRangeEnd = rangeEnd
 			i += 1
-		self.logger.debug("OffpeakStrategy.getPeakPeriod(%s) = %s", schedule, peakPeriods)
+		self.logger.debug("OffpeakStrategy.getPeakPeriod(%s) = %s",
+		                  schedule, peakPeriods)
 		return peakPeriods
 
 	def getOnPeriods(self, now, schedule):
@@ -230,11 +230,13 @@ class OffPeakStrategy(EnergyStrategy):
 		if missingTime<=TIMEDELTA_0:
 			return []
 		peakPeriods = self.getPeakPeriods(now, schedule)
+
 		# Choice when to start/stop: So choice the best periods
 		onPeriods = []
 		peakPeriods.sort(key=lambda x:x[3]) # Sort by cost
 		i = 0
-		while missingTime>TIMEDELTA_0:
+		while missingTime>TIMEDELTA_0 and i<len(peakPeriods):
+			self.logger.debug("Len(peakperiods)=%s, %s", len(peakPeriods), peakPeriods[i])
 			start, duration, end, _ = peakPeriods[i]
 			if duration>missingTime:
 				end = start + missingTime
@@ -244,6 +246,12 @@ class OffPeakStrategy(EnergyStrategy):
 			onPeriods.append([start, end]) # Sort by start time
 			i += 1
 		onPeriods.sort(key=lambda x:x[0])
+		if missingTime>TIMEDELTA_0:
+			end = onPeriods[len(onPeriods)-1][1]
+			end = end + missingTime
+			onPeriods[len(onPeriods)-1][1] = end
+			self.logger.warning("Pushing back last end (until %s) due to missing time (%d minutes).",
+			                    end, round(missingTime.seconds/60))
 		self.logger.debug("OffpeakStrategy.getOnPeriod(%s) = %s", schedule, onPeriods)
 		return onPeriods